MedicationService unit tests implementation

+++ src/main/java/com/openclassrooms/safety_net/service/MedicationService.java
@@ -3,7 +3,9 @@ package com.openclassrooms.safety_net.service;
 import com.openclassrooms.safety_net.model.Medication;
 import com.openclassrooms.safety_net.repository.MedicationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,16 +22,31 @@ public class MedicationService {
 	}
 
 	public List<Medication> getMedicationsOnlyWithoutMedicalRecordLink () {
-		Iterable<Medication> medications = medicationRepository.findAllergiesOnlyWithNoLinks();
+		Iterable<Medication> medications = medicationRepository.findMedicationsOnlyWithNoLinks();
 		return StreamSupport.stream(medications.spliterator(), false).toList();
 	}
 
-	public void deleteMedication (Medication medication) {
-		medicationRepository.delete(medication);
+	public void deleteMedication (Medication medication) throws ResponseStatusException {
+		if (medicationRepository.existsById(medication.getId())) {
+			medicationRepository.delete(medication);
+		} else {
+			throw new ResponseStatusException(
+					HttpStatus.NOT_FOUND,
+					"Medication %s whit dosage %dmg not found and can't be deleted.".formatted(
+							medication.getName(),
+							medication.getDosage()
+					)
+			);
+		}
 	}
 
 	public void deleteMedicationsWithoutMedicalRecordLink () {
 		List<Medication> medications = getMedicationsOnlyWithoutMedicalRecordLink();
-		medications.forEach(this::deleteMedication);
+		try {
+			medications.forEach(this::deleteMedication);
+		} catch (ResponseStatusException e) {
+			e.printStackTrace();
+		}
 	}
+
 }
